fix(FeatureMap): [TELFE-647] refactor mv
